+++ src/qtt/measurements/videomode.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Contains code to do live plotting 
+Contains code for the VideoMode tools
 
 """
 # %%
@@ -8,200 +8,153 @@ import time
 import datetime
 import threading
 import numpy as np
-import warnings
 import logging
-from scipy import ndimage
+import numbers
+
 import pyqtgraph
 
-import qtt
 from qcodes.instrument.parameter import Parameter
 from qcodes.utils.validators import Numbers
+import qtt
 from qtt.gui.live_plotting import livePlot
 from qtt.utilities.tools import connect_slot
 import qtpy.QtWidgets as QtWidgets
 import qtpy.QtCore
-from qtt.measurements.scans import makeDataset_sweep, makeDataset_sweep_2D
-from qtt.measurements.acquisition.interfaces import AcquisitionScopeInterface
+from qtt.measurements.videomode_processor import VideomodeSawtoothMeasurement
 
 # %%
 
 
-class videomode_callback:
+def add_sawtooth_videomode_processor(self, sweepparams, sweepranges, resolution, sample_rate, minstrument):
+    """ Add all required variables to the VideoMode for the VideomodeSawtoothMeasurement """
+    self.resolution = resolution
+    self.sample_rate = sample_rate
+    self.minstrumenthandle = minstrument[0]
+    channels = minstrument[1]
+    if isinstance(channels, int):
+        channels = [channels]
+    self.channels = channels
 
-    def __init__(self, station, waveform, Naverage, minstrument,
-                 diff_dir=None, resolution=None):
-        """ Create callback object for videmode data
+    self.videomode_processor = VideomodeSawtoothMeasurement(self.station)
 
-        Args:
-            station (QCoDeS station)
-            waveform
-            Naverage (int): number of average to take
-            minstrument (tuple): instrumentname, channel
-            diff_dir (list or (int, str)): differentiation modes for the data
-        """
-        self.station = station
-        self.waveform = waveform
-        self.Naverage = Naverage
-        self.minstrument = minstrument[0]
-        self.channel = minstrument[1]
-        self.channels = self.channel
-        self.measuresegment_arguments = {'mV_range': 5000}
+    sampling_frequency = self.videomode_processor.parse_instrument(self.minstrumenthandle, sample_rate)
 
-        if not isinstance(self.channels, list):
-            self.channels = [self.channels]
-
-        self.unique_channels = list(np.unique(self.channels))
-
-        # for 2D scans
-        self.resolution = resolution
-        self.diffsigma = 1
-        self.diff = True
-        self.diff_dir = diff_dir
-        self.smoothing = False
-        self.laplace = False
-
-    def __call__(self, verbose=0):
-        """ Callback function to read a single line of data from the device 
-
-        Returns:
-            data (list or array): either a list with length the number of channels or a numpy array with all data
-        """
-
-        minstrumenthandle = qtt.measurements.scans.get_instrument(self.minstrument)
-
-        data = qtt.measurements.scans.measuresegment(
-            self.waveform, self.Naverage, minstrumenthandle, self.unique_channels, **self.measuresegment_arguments)
-
-        if np.all(data == 0):
-            raise Exception('data returned contained only zeros, aborting')
-
-        dd = []
-        for ii, channel in enumerate(self.channels):
-            uchannelidx = self.unique_channels.index(channel)
-            data_processed = np.array(data[uchannelidx])
-
-            if self.diff_dir is not None:
-                if isinstance(self.diff_dir, (list, tuple)):
-                    diff_dir = self.diff_dir[ii]
-                else:
-                    diff_dir = self.diff_dir
-                data_processed = qtt.utilities.tools.diffImageSmooth(
-                    data_processed, dy=diff_dir, sigma=self.diffsigma)
-
-            if self.smoothing:
-                data_processed = qtt.algorithms.generic.smoothImage(
-                    data_processed)
-
-            if self.laplace:
-                data_processed = ndimage.filters.laplace(
-                    data_processed, mode='nearest')
-            dd.append(data_processed)
-        return dd
+    self.videomode_processor.set_scan_parameters(
+        {'sweepparams': sweepparams, 'sweepranges': sweepranges, 'minstrument': minstrument,
+         'resolution': self.resolution, 'sampling_frequency': sampling_frequency, 'channels': channels})
 
 
 # %%
 
 
 class VideoMode:
-    """ Controls the videomode tuning.
+    """ Controls the videomode tool.
+
+    The VideoMode tools allows for fast plotting of measurement results. The basic operation of the VideoMode consists
+    of the following stages:
+
+        i) Initialize. For example start a periodic waveform on the AWG
+        ii) Start the readout. This starts a loop with the following steps:
+            - Measure data
+            - Post-process data
+            - Plot data
+            The loop continues running in the background untill the user aborts the loop.
+        iii) Stop the readout. This stops the measure-process-plot loop
+        iv) Stop. This stops all activity (e.g. both the readout loop and and activity on the AWG)
+
 
     Attributes:
         station (qcodes station): contains all the information about the set-up
-        sweepparams (string, 1 x 2 list or dict): the parameter(s) to be swept
-        sweepranges (int or 1 x 2 list): the range(s) to be swept over (peak-to-peak range)
-        minstrument (int or tuple): the channel of the FPGA, or tuple (instrument, channel)
-        Naverage (int): the number of times the FPGA averages
-        resolution (1 x 2 list): for 2D the resolution
-        nplots (int or None): number of plots to show. must be equal to the number of channels in the minstrument argument
-        sample_rate (float): sample rate for acquisition device
-        crosshair (bool): enable crosshair
+        videomode_processor (VideoModeProcessor): class performing the measurements and post-processing
+        Naverage (Parameter): the number of times the raw measurement data should be averaged
     """
 
     videomode_class_index = 0
 
-    def __init__(self, station, sweepparams, sweepranges, minstrument, nplots=None, Naverage=10,
-                 resolution=[96, 96], sample_rate='default', diff_dir=None, verbose=1,
+    def __init__(self, station, sweepparams=None, sweepranges=None, minstrument=None, nplots=None, Naverage=10,
+                 resolution=(96, 96), sample_rate='default', diff_dir=None, verbose=1,
                  dorun=True, show_controls=True, add_ppt=True, crosshair=False, averaging=True, name=None,
-                 mouse_click_callback=None):
+                 mouse_click_callback=None, videomode_processor=None):
         """ Tool for fast acquisition of charge stability diagram
+
+        The acquisition and post-processing of data is performed by a VideoModeProcessor.
 
         The class assumes that the station has a station.digitizer and a station.awg.
 
         Args:
             station (QCoDeS station): reference for the instruments
             sweepparams (list or dict or str): parameters to sweep
+            sweepranges (list): list of sweep ranges
+            minstrument (object): specification of the measurement instrument
+            crosshair (bool): Enable crosshair in plots
+            videomode_processor (None or VideoModeProcessor): Processor to use for measurement and post-processing
             mouse_click_callback (None or function): if None then update scan position with callback
 
         """
         if VideoMode.videomode_class_index == 0:
             # create instance of QApplication
             _ = pyqtgraph.mkQApp()
-        VideoMode.videomode_class_index = VideoMode.videomode_class_index + 1  # increment index of VideoMode objects
+        VideoMode.videomode_class_index = VideoMode.videomode_class_index + 1
         self.videomode_index = VideoMode.videomode_class_index
 
         self.station = station
         self.verbose = verbose
         self.sweepparams = sweepparams
+        if isinstance(sweepranges, numbers.Number):
+            sweepranges = [sweepranges]
         self.sweepranges = sweepranges
-        self.virtual_awg = getattr(station, ' virtual_awg', None)
 
-        self.minstrumenthandle = minstrument[0]
-        self.channels = minstrument[1]
-        if isinstance(self.channels, int):
-            self.channels = [self.channels]
+        self.virtual_awg = getattr(station, ' virtual_awg', None)
 
         self.Naverage = Parameter('Naverage', get_cmd=self._get_Naverage,
                                   set_cmd=self._set_Naverage, vals=Numbers(1, 1023))
-        self._Naverage_val = Naverage
-        self.resolution = resolution
-        self.sample_rate = sample_rate
+
+        # set core VideoMode properties
         self.diff_dir = diff_dir
+        self.smoothing = 0
         self.datalock = threading.Lock()
         self.datafunction_result = None
         self.update_sleep = 1e-5
-
-        if name is None:
-            name = 'VideoMode %d' % self.videomode_index
-            if isinstance(self.sweepparams, (str, list)):
-                name += ': %s' % str(self.sweepparams)
-        self.name = name
-
-        # parse instrument
-        minstrumenthandle = qtt.measurements.scans.get_instrument(self.minstrumenthandle, station)
-
-        if isinstance(minstrumenthandle, AcquisitionScopeInterface):
-            if sample_rate != 'default':
-                minstrumenthandle.sample_rate = sample_rate
-            self.sampling_frequency = lambda: minstrumenthandle.sample_rate
-        elif minstrumenthandle.name in ['digitizer', 'm4i']:
-            if sample_rate == 'default':
-                self.sampling_frequency = minstrumenthandle.sample_rate
-            else:
-                minstrumenthandle.sample_rate(sample_rate)
-                self.sampling_frequency = station.digitizer.sample_rate
-        elif minstrumenthandle.name in ['ZIUHFLI', 'ziuhfli']:
-            self.sampling_frequency = minstrumenthandle.scope_samplingrate
-        else:
-            try:
-                minstrumenthandle = qtt.measurements.scans.get_instrument(
-                    minstrument, station)
-                self.sampling_frequency = minstrumenthandle.sample_rate
-            except:
-                raise Exception('no fpga or digitizer found')
-
-        # for addPPT and debugging
-        self.scanparams = {'sweepparams': sweepparams, 'sweepranges': sweepranges, 'minstrument': minstrument,
-                           'resolution': self.resolution, 'sampling_frequency': self.sampling_frequency()}
+        self.diffsigma = 1
+        self.diff = True
+        self.laplace = False
         self.idx = 0
         self.fps = qtt.pgeometry.fps_t(nn=6)
 
-        # Setup the GUI
-        if nplots is None:
-            nplots = len(self.channels)
+        if videomode_processor is None:
+            add_sawtooth_videomode_processor(self, sweepparams, sweepranges, resolution, sample_rate, minstrument)
+            # Setup the GUI
+            if nplots is None:
+                nplots = len(self.channels)
+        else:
+            self.videomode_processor = videomode_processor
+
+        self.set_videomode_name(name)
+
         self.nplots = nplots
         self.window_title = "%s: nplots %d" % (
             self.name, self.nplots)
 
+        self._create_gui(nplots, show_controls, add_ppt, crosshair, Naverage, averaging)
+
+        self.Naverage(Naverage)
+
+        if mouse_click_callback is None:
+            mouse_click_callback = self._update_position
+        for p in self.lp:
+            p.sigMouseClicked.connect(mouse_click_callback)
+
+        if dorun:
+            self.run()
+
+    def set_videomode_name(self, name):
+        """ Set the name for this instance of the tool """
+        if name is None:
+            name = 'VideoMode %d' % self.videomode_index
+            name = self.videomode_processor.extend_videomode_name(name)
+        self.name = name
+
+    def _create_gui(self, number_of_plots, show_controls, add_ppt, crosshair, Naverage, averaging):
         win = QtWidgets.QWidget()
         win.setWindowTitle(self.window_title)
         self.mainwin = win
@@ -234,10 +187,11 @@ class VideoMode:
         vertLayout.addLayout(self.plotLayout)
 
         self.lp = []
-        for ii in range(nplots):
-            lp = livePlot(None, self.station.gates,
+        plot_dimension = self.videomode_processor.scan_dimension()
+        for ii in range(number_of_plots):
+            lp = livePlot(self.videomode_processor, self.station.gates,
                           self.sweepparams, self.sweepranges, show_controls=False,
-                          plot_title=str(self.minstrumenthandle) + ' ' + str(self.channels[ii]))
+                          plot_title=self.videomode_processor.plot_title(ii), plot_dimension=plot_dimension)
             self.lp.append(lp)
             self.plotLayout.addWidget(self.lp[ii].win)
 
@@ -261,14 +215,6 @@ class VideoMode:
 
         self.enable_averaging_slot(averaging=averaging)
 
-        if mouse_click_callback is None:
-            mouse_click_callback = self._update_position
-        for p in self.lp:
-            p.sigMouseClicked.connect(mouse_click_callback)
-
-        if dorun:
-            self.run()
-
     def __repr__(self):
         s = '%s: %s at 0x%x' % (self.__class__.__name__, self.name, id(self))
         return s
@@ -280,42 +226,8 @@ class VideoMode:
             position (array): point with new coordinates
             verbose (int): verbosity level
         """
-        station = self.station
-
-        if verbose:
-            print('# %s: update position: %s' % (self.__class__.__name__, position,))
-
-        if self.scan_dimension() == 1:
-            delta = position[0]
-            if isinstance(self.scanparams['sweepparams'], str):
-                param = getattr(station.gates, self.scanparams['sweepparams'])
-                param.set(delta)
-                if verbose > 2:
-                    print('  set %s to %s' % (param, delta))
-                return
-        try:
-            for ii, parameter in enumerate(self.scanparams['sweepparams']):
-                delta = position[ii]
-                if verbose:
-                    print('param %s: delta %.3f' % (parameter, delta))
-
-                if isinstance(parameter, str):
-                    if parameter == 'gates_horz' or parameter == 'gates_vert':
-                        d = self.scanparams['sweepparams'][parameter]
-                        for g, f in d.items():
-                            if verbose >= 2:
-                                print('  %s: increment %s with %s' % (parameter, g, f * delta))
-                            param = getattr(station.gates, g)
-                            param.increment(f * delta)
-                    else:
-                        if verbose > 2:
-                            print('  set %s to %s' % (parameter, delta))
-                        param = getattr(station.gates, parameter)
-                        param.set(delta)
-                else:
-                    raise Exception('_update_position with parameter type %s not supported' % (type(parameter),))
-        except Exception as ex:
-            logging.exception(ex)
+        if isinstance(self.videomode_processor, VideomodeSawtoothMeasurement):
+            self.videomode_processor.update_position(position, verbose=verbose)
 
     def enable_averaging_slot(self, averaging=None, *args, **kwargs):
         """ Update the averaging mode of the widget """
@@ -340,36 +252,34 @@ class VideoMode:
 
         qtt.utilities.tools.addPPTslide(fig=self, title='VideoMode %s' % self.name,
                                         notes=self.station,
-                                        extranotes='date: %s' % (qtt.data.dateString(),) + '\n' + 'scanjob: ' + str(
-                                            self.scanparams) + '\n\n' + setting_notes)
+                                        extranotes='date: %s' % (
+                                            qtt.data.dateString(),) + '\n' + 'videomode_processor: ' + str(
+                                            self.videomode_processor.ppt_notes()) + '\n\n' + setting_notes)
         if isrunning:
             self.startreadout()
 
     def updatebg(self):
-        """ Update function for the widget 
+        """ Update function for the tool
 
-        Calls the datafunction() and update() function for all subplots
+        Calls the videomode_processor.measure() and videomode_processor.process() and updates the GUI
         """
         if self.idx % 10 == 0:
             logging.debug('%s: updatebg %d' %
                           (self.__class__.__name__, self.idx))
         self.idx = self.idx + 1
         self.fps.addtime(time.time())
-        if self.datafunction is not None:
-            try:
-                dd = self.datafunction()
-                self.datafunction_result = dd
-                for ii, d in enumerate(dd):
-                    self.lp[ii].update(data=d, processevents=False)
-                pyqtgraph.mkQApp().processEvents()
-            except Exception as e:
-                logging.exception(e)
-                print('%s: Exception in updatebg, stopping readout' %
-                      self.__class__.__name__)
-                self.stopreadout()
-        else:
+        try:
+            measured_data = self.videomode_processor.measure(self)
+            datafunction_result = self.videomode_processor.process(measured_data, self)
+            self.datafunction_result = datafunction_result
+            for ii, d in enumerate(datafunction_result):
+                self.lp[ii].update(data=d, processevents=False)
+            pyqtgraph.mkQApp().processEvents()
+        except Exception as ex:
+            logging.exception(ex)
+            print('%s: Exception in updatebg, stopping readout' %
+                  self.__class__.__name__)
             self.stopreadout()
-            dd = None
 
         self.mainwin.setWindowTitle(
             self.window_title + ' %.1f [fps]' % self.fps.framerate())
@@ -379,10 +289,12 @@ class VideoMode:
         time.sleep(self.update_sleep)
 
     def is_running(self):
+        """ Return True if the readout loop is running """
         return self.timer.isActive()
 
     def startreadout(self, callback=None, rate=30, maxidx=None):
-        """
+        """ Start the readout loop
+
         Args:
             rate (float): sample rate in ms
 
@@ -396,6 +308,7 @@ class VideoMode:
             print('%s: start readout' % (self.__class__.__name__,))
 
     def stopreadout(self):
+        """ Stop the readout loop """
         if self.verbose:
             print('%s: stop readout' % (self.__class__.__name__,))
         self.timer.stop()
@@ -415,9 +328,11 @@ class VideoMode:
         return box
 
     def close(self):
+        """ Stop the videomode and close the GUI"""
+        self.stop()
         self.mainwin.close()
 
-    def get_dataset(self, run=False):
+    def get_dataset(self):
         """ Return latest recorded dataset
 
         Returns:
@@ -425,179 +340,60 @@ class VideoMode:
 
         """
         with self.datalock:
-            if run:
-                warnings.warn('not supported')
-                # [l.datafunction_result for l in self.lp]
-                data = self.datafunction_result
-                data = np.array(data)
-            else:
-                data = self.datafunction_result
-                if data is not None:
-                    data = np.array(data)
+            data = self.datafunction_result
             if data is not None:
-                self.alldata = self.makeDataset(data, Naverage=None)
+                data = np.array(data)
+                self.alldata = self._makeDataset(data, Naverage=None)
+            else:
+                self.alldata = None
             return self.alldata
 
-    def scan_dimension(self):
-        if isinstance(self.sweepranges, (float, int)):
-            return 1
-        elif isinstance(self.sweepranges, list):
-            if len(self.sweepranges) == 2:
-                return 2
-            elif len(self.sweepranges) == 1:
-                return 1
-            else:
-                raise Exception('scan dimension not supported')
-        else:
-            return -1
+    def initialize(self):
+        """ Initialize the videomode tool for the readout loop """
+        self.videomode_processor.initialize(self)
 
     def run(self, start_readout=True):
-        """ Programs the AWG, starts the read-out and the plotting. """
+        """ Initialize the tool and start the readout loop """
         if self.verbose:
             print('%s: run ' % (self.__class__.__name__,))
-        scan_dimensions = self.scan_dimension()
-        virtual_awg = getattr(self.station, 'virtual_awg', None)
-        awg = getattr(self.station, 'awg', None)
 
-        if scan_dimensions == 1:
-            self.__run_1d_scan(awg, virtual_awg)
-        elif scan_dimensions == 2:
-            self.__run_2d_scan(awg, virtual_awg)
-        else:
-            raise Exception('type of scan not supported')
+        self.initialize()
 
         if start_readout:
             if self.verbose:
                 print('%s: run: startreadout' % (self.__class__.__name__,))
-            if isinstance(self.minstrumenthandle, AcquisitionScopeInterface):
-                if scan_dimensions == 1:
-                    self.minstrumenthandle.period = 1e-3
-                elif scan_dimensions == 2:
-                    self.minstrumenthandle.number_of_samples = np.prod(self.resolution)
-                scan_channels = self.scanparams['minstrument'][1]
-                channel_attributes = self.scanparams['minstrument'][2]
-                self.minstrumenthandle.enabled_channels = tuple(scan_channels)
-                for channel, attribute in zip(scan_channels, channel_attributes):
-                    self.minstrumenthandle.set_input_signal(channel, attribute)
-                self.minstrumenthandle.start_acquisition()
             self.startreadout()
 
-    def __run_1d_scan(self, awg, virtual_awg, period=1e-3):
-        if virtual_awg:
-            if not isinstance(self.sweepparams, (str, dict)):
-                raise Exception('argument sweepparams of type %s not supported' % type(self.sweepparams))
-            sweep_range = self.sweepranges
-            gates = self.sweepparams if isinstance(self.sweepparams, dict) else {self.sweepparams: 1}
-            waveform = virtual_awg.sweep_gates(gates, sweep_range, period)
-            virtual_awg.enable_outputs(list(gates.keys()))
-            virtual_awg.run()
-        else:
-            if type(self.sweepparams) is str:
-                waveform, _ = awg.sweep_gate(self.sweepparams, self.sweepranges, period=period)
-            elif type(self.sweepparams) is dict:
-                waveform, _ = awg.sweep_gate_virt(self.sweepparams, self.sweepranges, period=period)
-            else:
-                raise Exception('argument sweepparams of type %s not supported' % type(self.sweepparams))
-        self.datafunction = videomode_callback(self.station, waveform, self.Naverage.get(),
-                                               minstrument=(self.minstrumenthandle, self.channels))
-
-    def __run_2d_scan(self, awg, virtual_awg, period=None):
-        if virtual_awg:
-            sweep_ranges = [i for i in self.sweepranges]
-            if isinstance(self.sweepparams[0], dict):
-                gates = self.sweepparams
-            else:
-                # convert list of strings to dict format
-                gates = self.sweepparams
-                gates = [dict([(gate, 1)]) for gate in gates]
-
-            if period is None:
-                sampling_frequency = qtt.measurements.scans.get_sampling_frequency(self.minstrumenthandle)
-                base_period = 1. / sampling_frequency
-                total_period = base_period * np.prod(self.resolution)
-            else:
-                total_period = period
-                warnings.warn('code untested')
-            waveform = virtual_awg.sweep_gates_2d(gates, sweep_ranges, total_period, self.resolution)
-            keys = [list(item.keys())[0] for item in gates]
-            virtual_awg.enable_outputs(keys)
-            virtual_awg.run()
-        else:
-            if isinstance(self.sweepparams, list):
-                waveform, _ = awg.sweep_2D(self.sampling_frequency.get(), self.sweepparams,
-                                           self.sweepranges, self.resolution)
-            elif isinstance(self.sweepparams, dict):
-                waveform, _ = awg.sweep_2D_virt(self.sampling_frequency.get(), self.sweepparams[
-                    'gates_horz'], self.sweepparams['gates_vert'], self.sweepranges, self.resolution)
-            else:
-                raise Exception('argument sweepparams of type %s not supported' % type(self.sweepparams))
-            if self.verbose:
-                print('%s: 2d scan, define callback ' % (self.__class__.__name__,))
-        self.datafunction = videomode_callback(self.station, waveform, self.Naverage.get(),
-                                               minstrument=(self.minstrumenthandle, self.channels),
-                                               resolution=self.resolution, diff_dir=self.diff_dir)
-
     def stop(self):
-        """ Stops the plotting, AWG(s) and if available RF. """
+        """ Stops the readout loop and the input signals """
         self.stopreadout()
-        if self.station is not None:
-            if hasattr(self.station, 'awg'):
-                self.station.awg.stop()
-            if hasattr(self.station, 'RF'):
-                self.station.RF.off()
-            if hasattr(self.station, 'virtual_awg'):
-                self.station.virtual_awg.stop()
-                if isinstance(self.sweepparams[0], dict):
-                    keys = [list(item.keys())[0] for item in self.sweepparams]
-                    self.station.virtual_awg.disable_outputs(keys)
-                elif isinstance(self.sweepparams, str):
-                    self.station.virtual_awg.disable_outputs([self.sweepparams])
+        self.stop_videomode()
 
-        if isinstance(self.minstrumenthandle, AcquisitionScopeInterface):
-            self.minstrumenthandle.stop_acquisition()
+    def stop_videomode(self):
+        if self.verbose:
+            print('%s: stop_videomode ' % (self.__class__.__name__,))
 
+        self.videomode_processor.stop()
 
     def single(self):
-        """Do a single scan with a lot averaging.
+        """ Do a single scan with a lot averaging.
 
         Note: this does not yet support the usage of linear combinations of gates (a.k.a. virtual gates).
         """
         raise Exception('not implemented')
 
     def crosshair(self, *args, **kwargs):
+        """ Enable or disable a crosshair in the plotting windows """
         for l in self.lp:
             l.crosshair(*args, **kwargs)
 
-    def makeDataset(self, data, Naverage=None):
-        """ Helper function """
+    def _makeDataset(self, data, Naverage=None):
+        """ Create datasets from the processed data """
         metadata = {'scantime': str(datetime.datetime.now(
         )), 'station': self.station.snapshot(), 'allgatevalues': self.station.gates.allvalues()}
         metadata['Naverage'] = Naverage
-        if hasattr(self.datafunction, 'diff_dir'):
-            metadata['diff_dir'] = self.datafunction.diff_dir
 
-        if data.ndim == 2:
-            if (data.shape[0] > 1):
-                raise Exception('not yet implemented')
-            data = data[0]
-            alldata, _ = makeDataset_sweep(data, self.sweepparams, self.sweepranges,
-                                           gates=self.station.gates, loc_record={'label': 'videomode_1d_single'})
-            alldata.metadata = metadata
-        elif data.ndim == 3:
-            if (data.shape[0] > 1):
-                warnings.warn('getting dataset for multiple dimensions not yet tested')
-
-            import copy
-            alldata = [None] * len(data)
-            for jj in range(len(data)):
-                datax = data[jj]
-                alldatax, _ = makeDataset_sweep_2D(datax, self.station.gates, self.sweepparams, self.sweepranges,
-                                                   loc_record={
-                                                       'label': 'videomode_2d_single'})
-                alldatax.metadata = copy.copy(metadata)
-                alldata[jj] = alldatax
-        else:
-            raise Exception('makeDataset: data.ndim %d' % data.ndim)
+        alldata = self.videomode_processor.create_dataset(data, metadata)
         return alldata
 
     def _get_Naverage(self):
@@ -605,7 +401,6 @@ class VideoMode:
 
     def _set_Naverage(self, value):
         self._Naverage_val = value
-        self.datafunction.Naverage = value
         self.box.setValue(value)
 
     @staticmethod
@@ -631,49 +426,22 @@ class VideoMode:
             v.stopreadout()
 
 
-def guitest_videomode():
-    import pyqtgraph
-    _ = pyqtgraph.mkQApp()
-    from qtt.instrument_drivers.simulation_instruments import SimulationDigitizer
-    from qtt.instrument_drivers.simulation_instruments import SimulationAWG
-
-    assert(VideoMode.get_instance(-1) is None)
-
-    import qtt.simulation.virtual_dot_array
-    station = qtt.simulation.virtual_dot_array.initialize()
-
-    digitizer = SimulationDigitizer(
-        qtt.measurements.scans.instrumentName('sdigitizer'), model=station.model)
-    station.add_component(digitizer)
-
-    station.awg = SimulationAWG(qtt.measurements.scans.instrumentName('vawg'))
-    station.add_component(station.awg)
-
-    sweepparams = {'gates_horz': {'P1': 1}, 'gates_vert': {'P2': 1}}
-    minstrument = (digitizer.name, [0])
-
-    vm = VideoMode(station, sweepparams, sweepranges=[120] * 2,
-                   minstrument=minstrument, resolution=[12] * 2, Naverage=2)
-    vm.stop()
-    vm.close()
-
-    all = VideoMode.all_instances(verbose=0)
-    assert(vm in all)
-
-
 # %% Testing
 
 
 if __name__ == '__main__':
-    guitest_videomode()
-
     import qtt.simulation.virtual_dot_array
-    station = qtt.simulation.virtual_dot_array.initialize()
-    gates = station.gates
+    import qtt.instrument_drivers.simulation_instruments
+    from importlib import reload
 
+    reload(qtt.instrument_drivers.simulation_instruments)
+    reload(qtt.measurements.scans)
     from qtt.instrument_drivers.simulation_instruments import SimulationDigitizer
     from qtt.instrument_drivers.simulation_instruments import SimulationAWG
-    #from qtt.gui.live_plotting import *
+    from qtt.measurements.videomode_processor import DummyVideoModeProcessor
+
+    station = qtt.simulation.virtual_dot_array.initialize()
+    gates = station.gates
 
     pv = qtt.createParameterWidget([gates])  # type: ignore
 
@@ -682,10 +450,24 @@ if __name__ == '__main__':
 
     digitizer = SimulationDigitizer(
         qtt.measurements.scans.instrumentName('sdigitizer'), model=station.model)
-    station.components[digitizer.name] = digitizer
+    station.add_component(digitizer)
 
     station.awg = SimulationAWG(qtt.measurements.scans.instrumentName('vawg'))
-    station.components[station.awg.name] = station.awg
+    station.add_component(station.awg)
+
+    dummy_processor = DummyVideoModeProcessor(station)
+    vm = VideoMode(station, Naverage=25, diff_dir=None, verbose=1,
+                   nplots=1, dorun=False, videomode_processor=dummy_processor)
+    vm.updatebg()
+
+    self = vm
+
+    if 0:
+        sweepparams = [{'P1': 1}, {'P2': 1}]
+        minstrument = (digitizer.name, [0])
+        vm = VideoMode(station, sweepparams, sweepranges=[120] * 2,
+                       minstrument=minstrument, resolution=[12] * 2, Naverage=2)
+        self = vm
 
     if 1:
         sweepparams = ['P1', 'P2']
@@ -700,46 +482,19 @@ if __name__ == '__main__':
     station.model.sdnoise = .1
     vm = VideoMode(station, sweepparams, sweepranges, minstrument, Naverage=25,
                    resolution=resolution, sample_rate='default', diff_dir=None,
-                   verbose=1, nplots=None, dorun=True)
+                   verbose=1, nplots=None, dorun=False)
+    self = vm
+    self.initialize()
+    measured_data = self.videomode_processor.measure(vm)
+    dd = self.videomode_processor.process(measured_data, vm)
+
+    vm.updatebg()
+    vm.run()
+    vm.get_dataset()
 
     vm.setGeometry(1310, 100, 800, 800)
+    self = vm
 
     vm.stopreadout()
     vm.updatebg()
     vm.stop()
-
-    # %% Test MultiTracePlot
-    app = pyqtgraph.mkQApp()
-    waveform, ix = station.awg.sweep_gate('P1', 50, 1e-3)
-    nplots = 3
-    ncurves = 2
-
-    def read_trace_dummy():
-        data = qtt.measurements.scans.measuresegment(
-            waveform, Naverage=1, minstrhandle=station.sdigitizer.name, read_ch=[1, 2])
-        dd = [data] * nplots
-        xd = np.linspace(-waveform['sweeprange'] / 2,
-                         waveform['sweeprange'] / 2, data[0].size)
-        xdata = [xd] * nplots
-        return xdata, dd
-
-    xd, yd = read_trace_dummy()
-
-    # %%
-    import qtt.measurements.ttrace
-    from qtt.measurements.ttrace import MultiTracePlot
-
-    mt = MultiTracePlot(nplots=nplots, ncurves=ncurves)
-    mt.win.setGeometry(1400, 40, 500, 500)
-    mt.add_verticals()
-
-    def callback():
-        xdata, ydata = read_trace_dummy()
-        mt.plot_curves(xdata, ydata)
-        app.processEvents()
-
-    mt.startreadout(callback=callback)
-    mt.updatefunction()
-
-    mt.get_dataset()
-    mt.stopreadout()